Back from the long detour.. We are ready to try to fix all the combined-units problems we had

+++ EngineeringUnits/CombinedUnits/Area/Area.cs
@@ -21,9 +21,9 @@ namespace EngineeringUnits
             SetLocalValue(value);
         }
 
-        public decimal As(AreaUnit SquaredlengthUnit)
+        public double As(AreaUnit SquaredlengthUnit)
         {
-            return ToTheOutSide(SquaredlengthUnit.Unit);
+            return (double)ToTheOutSide(SquaredlengthUnit.Unit);
         }
 
         //Every units needs this
